+++ multiplatform-locale-all/src/commonMain/kotlin/com/vanniktech/locale/all/iosworkaround.kt
@@ -1,0 +1,5 @@
+@file:Suppress("standard:no-empty-file")
+
+package com.vanniktech.locale.all
+
+// Empty file is required for iOS, otherwise it fails.