+++ multiplatform-locale/src/commonMain/kotlin/com/vanniktech/locale/Language.kt
@@ -5,306 +5,314 @@ enum class Language(
   val code: String,
   /** [HEBREW] & [INDONESIAN] changed their identifier. We are backwards compatible and support both. */
   val legacyCode: String? = null,
-  /** Countries, where this language is officially spoken in. */
-  val officialCountries: List<Country>,
-  /** Countries, where this language is spoken in. */
-  val otherCountries: List<Country>,
+  /** Territories, where this [Language] is officially spoken in. */
+  val officialTerritories: List<Territory>,
+  /** Territories, where this [Language] is spoken in. */
+  val otherTerritories: List<Territory>,
 ) {
   ARABIC(
     code = "ar",
-    officialCountries = listOf(Country.SAUDI_ARABIA, Country.ALGERIA, Country.BAHRAIN, Country.MALI, Country.NIGER, Country.SENEGAL, Country.CHAD, Country.COMOROS, Country.DJIBOUTI, Country.EGYPT, Country.IRAQ, Country.JORDAN, Country.KUWAIT, Country.LEBANON, Country.LIBYA, Country.MAURITANIA, Country.MOROCCO, Country.OMAN, Country.QATAR, Country.SOMALIA, Country.SUDAN, Country.SYRIA, Country.TUNISIA, Country.UNITED_ARAB_EMIRATES, Country.YEMEN),
-    otherCountries = listOf(Country.PALESTINE, Country.SOMALIA, Country.SAHRAWI_ARAB_DEMOCRATIC_REPUBLIC, Country.TANZANIA),
+    officialTerritories = listOf(Country.SAUDI_ARABIA, Country.ALGERIA, Country.BAHRAIN, Country.MALI, Country.NIGER, Country.SENEGAL, Country.CHAD, Country.COMOROS, Country.DJIBOUTI, Country.EGYPT, Country.IRAQ, Country.JORDAN, Country.KUWAIT, Country.LEBANON, Country.LIBYA, Country.MAURITANIA, Country.MOROCCO, Country.OMAN, Country.QATAR, Country.SOMALIA, Country.SUDAN, Country.SYRIA, Country.TUNISIA, Country.UNITED_ARAB_EMIRATES, Country.YEMEN),
+    otherTerritories = listOf(Country.PALESTINE, Country.SOMALIA, Country.SAHRAWI_ARAB_DEMOCRATIC_REPUBLIC, Country.TANZANIA),
   ),
   AMHARIC(
     code = "am",
-    officialCountries = listOf(Country.ETHIOPIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.ETHIOPIA),
+    otherTerritories = emptyList(),
   ),
   ARMENIAN(
     code = "hy",
-    officialCountries = listOf(Country.ARMENIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.ARMENIA),
+    otherTerritories = emptyList(),
   ),
   AZERBAIJANI(
     code = "az",
-    officialCountries = listOf(Country.AZERBAIJAN),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.AZERBAIJAN),
+    otherTerritories = emptyList(),
   ),
   BELARUSIAN(
     code = "be",
-    officialCountries = listOf(Country.BELARUS),
-    otherCountries = listOf(Country.RUSSIA, Country.LITHUANIA, Country.LATVIA, Country.POLAND, Country.UKRAINE),
+    officialTerritories = listOf(Country.BELARUS),
+    otherTerritories = listOf(Country.RUSSIA, Country.LITHUANIA, Country.LATVIA, Country.POLAND, Country.UKRAINE),
   ),
   BOKMAL(
     code = "nb",
-    officialCountries = listOf(Country.NORWAY),
-    otherCountries = listOf(),
+    officialTerritories = listOf(Country.NORWAY),
+    otherTerritories = listOf(),
   ),
   BULGARIAN(
     code = "bg",
-    officialCountries = listOf(Country.BULGARIA),
-    otherCountries = listOf(Country.GREECE, Country.ROMANIA, Country.MOLDOVA, Country.UKRAINE),
+    officialTerritories = listOf(Country.BULGARIA),
+    otherTerritories = listOf(Country.GREECE, Country.ROMANIA, Country.MOLDOVA, Country.UKRAINE),
   ),
   BURMESE(
     code = "my",
-    officialCountries = listOf(Country.MYANMAR),
-    otherCountries = listOf(),
+    officialTerritories = listOf(Country.MYANMAR),
+    otherTerritories = listOf(),
   ),
   CENTRAL_KMHER(
     code = "km",
-    officialCountries = listOf(Country.CAMBODIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.CAMBODIA),
+    otherTerritories = emptyList(),
   ),
   CHINESE(
     code = "zh",
-    officialCountries = listOf(Country.CHINA, Country.TAIWAN, Country.HONG_KONG, Country.SINGAPORE),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.CHINA, Country.TAIWAN, Country.HONG_KONG, Country.SINGAPORE, Country.MACAO),
+    otherTerritories = emptyList(),
   ),
   CATALAN(
     code = "ca",
-    officialCountries = listOf(Country.SPAIN),
-    otherCountries = listOf(Country.ANDORRA),
+    officialTerritories = listOf(Country.SPAIN),
+    otherTerritories = listOf(Country.ANDORRA),
   ),
   CZECH(
     code = "cs",
-    officialCountries = listOf(Country.CZECHIA),
-    otherCountries = listOf(Country.POLAND, Country.SERBIA, Country.UKRAINE, Country.SLOVAKIA, Country.BULGARIA, Country.AUSTRIA),
+    officialTerritories = listOf(Country.CZECHIA),
+    otherTerritories = listOf(Country.POLAND, Country.SERBIA, Country.UKRAINE, Country.SLOVAKIA, Country.BULGARIA, Country.AUSTRIA),
   ),
   CROATIAN(
     code = "hr",
-    officialCountries = listOf(Country.CROATIA),
-    otherCountries = listOf(Country.AUSTRIA, Country.ITALY, Country.SERBIA),
+    officialTerritories = listOf(Country.CROATIA),
+    otherTerritories = listOf(Country.AUSTRIA, Country.ITALY, Country.SERBIA),
   ),
   DANISH(
     code = "da",
-    officialCountries = listOf(Country.DENMARK, Country.GREENLAND, Country.FAROE_ISLANDS),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.DENMARK, Country.GREENLAND, Country.FAROE_ISLANDS),
+    otherTerritories = emptyList(),
   ),
   DUTCH(
     code = "nl",
-    officialCountries = listOf(Country.NETHERLANDS, Country.BELGIUM, Country.SURINAME, Country.ARUBA, Country.CURACAO, Country.SINT_MAARTEN),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.NETHERLANDS, Country.BELGIUM, Country.SURINAME, Country.ARUBA, Country.CURACAO, Country.SINT_MAARTEN),
+    otherTerritories = emptyList(),
   ),
   ENGLISH(
     code = "en",
-    officialCountries = listOf(Country.USA, Country.CANADA, Country.AUSTRALIA, Country.ENGLAND, Country.NEW_ZEALAND, Country.CHRISTMAS_ISLAND, Country.COCOS_ISLANDS, Country.TOKELAU, Country.BRITISH_INDIAN_OCEAN_TERRITORY, Country.GUERNSEY, Country.MONTSERRAT),
-    otherCountries = listOf(Country.BAHRAIN, Country.BANGLADESH, Country.BRUNEI, Country.CYPRUS, Country.ERITREA, Country.ETHIOPIA, Country.GREENLAND, Country.ICELAND, Country.ISRAEL, Country.JORDAN, Country.KUWAIT, Country.MALAYSIA, Country.MALDIVES, Country.MYANMAR, Country.QATAR, Country.SRI_LANKA, Country.UNITED_ARAB_EMIRATES),
+    officialTerritories = listOf(Country.USA, Country.CANADA, Country.AUSTRALIA, Country.ENGLAND, Country.NEW_ZEALAND, Country.CHRISTMAS_ISLAND, Country.COCOS_ISLANDS, Country.TOKELAU, Country.BRITISH_INDIAN_OCEAN_TERRITORY, Country.GUERNSEY, Country.MONTSERRAT),
+    otherTerritories = listOf(Country.BAHRAIN, Country.BANGLADESH, Country.BRUNEI, Country.CYPRUS, Country.ERITREA, Country.ETHIOPIA, Country.GREENLAND, Country.ICELAND, Country.ISRAEL, Country.JORDAN, Country.KUWAIT, Country.MALAYSIA, Country.MALDIVES, Country.MYANMAR, Country.QATAR, Country.SRI_LANKA, Country.UNITED_ARAB_EMIRATES),
   ),
   ESTONIAN(
     code = "et",
-    officialCountries = listOf(Country.ESTONIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.ESTONIA),
+    otherTerritories = emptyList(),
   ),
   FARSI(
     code = "fa",
-    officialCountries = listOf(Country.IRAN),
-    otherCountries = listOf(Country.AFGHANISTAN, Country.BAHRAIN, Country.IRAQ, Country.OMAN, Country.YEMEN, Country.UNITED_ARAB_EMIRATES),
+    officialTerritories = listOf(Country.IRAN),
+    otherTerritories = listOf(Country.AFGHANISTAN, Country.BAHRAIN, Country.IRAQ, Country.OMAN, Country.YEMEN, Country.UNITED_ARAB_EMIRATES),
+  ),
+  FILIPINO(
+    code = "fil",
+    officialTerritories = listOf(Country.PHILIPPINES),
+    otherTerritories = emptyList(),
   ),
   FINNISH(
     code = "fi",
-    officialCountries = listOf(Country.FINLAND, Country.SWEDEN),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.FINLAND, Country.SWEDEN),
+    otherTerritories = emptyList(),
   ),
   FRENCH(
     code = "fr",
-    officialCountries = listOf(Country.FRANCE, Country.CANADA, Country.BELGIUM, Country.BENIN, Country.BURKINA_FASO, Country.BURUNDI, Country.CAMEROON, Country.CHAD, Country.COTE_DIVOIRE, Country.DEMOCRATIC_REPUBLIC_OF_CONGO, Country.DJIBOUTI, Country.EQUATORIAL_GUINEA, Country.HAITI, Country.LUXEMBOURG, Country.MADAGASCAR, Country.MALI, Country.MONACO, Country.NIGER, Country.RWANDA, Country.SENEGAL, Country.SEYCHELLES, Country.SWITZERLAND, Country.TOGO, Country.VANUATU),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.FRANCE, Country.CANADA, Country.BELGIUM, Country.BENIN, Country.BURKINA_FASO, Country.BURUNDI, Country.CAMEROON, Country.CHAD, Country.COTE_DIVOIRE, Country.DEMOCRATIC_REPUBLIC_OF_CONGO, Country.DJIBOUTI, Country.EQUATORIAL_GUINEA, Country.HAITI, Country.LUXEMBOURG, Country.MADAGASCAR, Country.MALI, Country.MONACO, Country.NIGER, Country.RWANDA, Country.SENEGAL, Country.SEYCHELLES, Country.SWITZERLAND, Country.TOGO, Country.VANUATU),
+    otherTerritories = emptyList(),
   ),
   GERMAN(
     code = "de",
-    officialCountries = listOf(Country.GERMANY, Country.AUSTRIA, Country.BELGIUM, Country.LUXEMBOURG, Country.SWITZERLAND, Country.LIECHTENSTEIN),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.GERMANY, Country.AUSTRIA, Country.BELGIUM, Country.LUXEMBOURG, Country.SWITZERLAND, Country.LIECHTENSTEIN),
+    otherTerritories = emptyList(),
   ),
   GREEK(
     code = "el",
-    officialCountries = listOf(Country.GREECE, Country.CYPRUS),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.GREECE, Country.CYPRUS),
+    otherTerritories = emptyList(),
   ),
   GUJARATI(
     code = "gu",
-    officialCountries = listOf(Country.INDIA, Country.SRI_LANKA, Country.SINGAPORE),
-    otherCountries = listOf(Country.MALAYSIA, Country.MAURITIUS, Country.FIJI, Country.SOUTH_AFRICA),
+    officialTerritories = listOf(Country.INDIA, Country.SRI_LANKA, Country.SINGAPORE),
+    otherTerritories = listOf(Country.MALAYSIA, Country.MAURITIUS, Country.FIJI, Country.SOUTH_AFRICA),
   ),
   HEBREW(
     code = "he",
     legacyCode = "iw",
-    officialCountries = listOf(Country.ISRAEL),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.ISRAEL),
+    otherTerritories = emptyList(),
+  ),
+  ICELANDIC(
+    code = "is",
+    officialTerritories = listOf(Country.ICELAND),
+    otherTerritories = emptyList(),
   ),
   HINDI(
     code = "hi",
-    officialCountries = listOf(Country.INDIA),
-    otherCountries = listOf(Country.MAURITIUS, Country.FIJI, Country.SURINAME, Country.GUYANA, Country.TRINIDAD_AND_TOBAGO, Country.NEPAL),
+    officialTerritories = listOf(Country.INDIA),
+    otherTerritories = listOf(Country.MAURITIUS, Country.FIJI, Country.SURINAME, Country.GUYANA, Country.TRINIDAD_AND_TOBAGO, Country.NEPAL),
   ),
   HUNGARIAN(
     code = "hu",
-    officialCountries = listOf(Country.HUNGARY),
-    otherCountries = listOf(Country.ROMANIA, Country.SLOVAKIA, Country.SERBIA, Country.AUSTRIA, Country.SLOVENIA),
+    officialTerritories = listOf(Country.HUNGARY),
+    otherTerritories = listOf(Country.ROMANIA, Country.SLOVAKIA, Country.SERBIA, Country.AUSTRIA, Country.SLOVENIA),
   ),
   INDONESIAN(
     code = "id",
     legacyCode = "in",
-    officialCountries = listOf(Country.INDONESIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.INDONESIA),
+    otherTerritories = emptyList(),
   ),
   IRISH(
     code = "ga",
-    officialCountries = listOf(Country.IRELAND),
-    otherCountries = listOf(),
+    officialTerritories = listOf(Country.IRELAND),
+    otherTerritories = listOf(),
   ),
   ITALIAN(
     code = "it",
-    officialCountries = listOf(Country.ITALY, Country.SAN_MARINO, Country.SWITZERLAND, Country.VATICAN_CITY),
-    otherCountries = listOf(Country.CROATIA, Country.SLOVENIA, Country.AUSTRIA),
+    officialTerritories = listOf(Country.ITALY, Country.SAN_MARINO, Country.SWITZERLAND, Country.VATICAN_CITY),
+    otherTerritories = listOf(Country.CROATIA, Country.SLOVENIA, Country.AUSTRIA),
   ),
   JAPANESE(
     code = "ja",
-    officialCountries = listOf(Country.JAPAN),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.JAPAN),
+    otherTerritories = emptyList(),
   ),
   KOREAN(
     code = "ko",
-    officialCountries = listOf(Country.SOUTH_KOREA, Country.NORTH_KOREA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.SOUTH_KOREA, Country.NORTH_KOREA),
+    otherTerritories = emptyList(),
   ),
   LATVIAN(
     code = "lv",
-    officialCountries = listOf(Country.LATVIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.LATVIA),
+    otherTerritories = emptyList(),
   ),
   LITHUANIAN(
     code = "lt",
-    officialCountries = listOf(Country.LITHUANIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.LITHUANIA),
+    otherTerritories = emptyList(),
   ),
   MALAY(
     code = "ms",
-    officialCountries = listOf(Country.MALAYSIA, Country.BRUNEI, Country.INDONESIA, Country.SINGAPORE),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.MALAYSIA, Country.BRUNEI, Country.INDONESIA, Country.SINGAPORE),
+    otherTerritories = emptyList(),
   ),
   MALAYALAM(
     code = "ml",
-    officialCountries = listOf(Country.INDIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.INDIA),
+    otherTerritories = emptyList(),
   ),
   MARATHI(
     code = "mr",
-    officialCountries = listOf(Country.INDIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.INDIA),
+    otherTerritories = emptyList(),
   ),
   NORWEGIAN(
     code = "no",
-    officialCountries = listOf(Country.NORWAY),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.NORWAY),
+    otherTerritories = emptyList(),
   ),
   POLISH(
     code = "pl",
-    officialCountries = listOf(Country.POLAND),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.POLAND),
+    otherTerritories = emptyList(),
   ),
   PORTUGUESE(
     code = "pt",
-    officialCountries = listOf(Country.PORTUGAL, Country.BRAZIL, Country.ANGOLA, Country.MOZAMBIQUE),
-    otherCountries = listOf(Country.GUINEA_BISSAU, Country.CAPE_VERDE, Country.SAO_TOME_AND_PRINCE, Country.MACAO),
+    officialTerritories = listOf(Country.PORTUGAL, Country.BRAZIL, Country.ANGOLA, Country.MOZAMBIQUE),
+    otherTerritories = listOf(Country.GUINEA_BISSAU, Country.CAPE_VERDE, Country.SAO_TOME_AND_PRINCE, Country.MACAO),
   ),
   ROMANIAN(
     code = "ro",
-    officialCountries = listOf(Country.ROMANIA, Country.MOLDOVA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.ROMANIA, Country.MOLDOVA),
+    otherTerritories = emptyList(),
   ),
   RUSSIAN(
     code = "ru",
-    officialCountries = listOf(Country.RUSSIA, Country.BELARUS, Country.KAZAKHSTAN, Country.KYRGYZSTAN),
-    otherCountries = listOf(Country.AZERBAIJAN, Country.ESTONIA, Country.GEORGIA, Country.LATVIA, Country.UKRAINE, Country.LITHUANIA, Country.MOLDOVA, Country.TAJIKISTAN, Country.TURKMENISTAN, Country.UZBEKISTAN, Country.ISRAEL),
+    officialTerritories = listOf(Country.RUSSIA, Country.BELARUS, Country.KAZAKHSTAN, Country.KYRGYZSTAN),
+    otherTerritories = listOf(Country.AZERBAIJAN, Country.ESTONIA, Country.GEORGIA, Country.LATVIA, Country.UKRAINE, Country.LITHUANIA, Country.MOLDOVA, Country.TAJIKISTAN, Country.TURKMENISTAN, Country.UZBEKISTAN, Country.ISRAEL),
   ),
   SERBIAN(
     code = "sr",
-    officialCountries = listOf(Country.SERBIA),
-    otherCountries = listOf(Country.CROATIA, Country.BOSNIA_AND_HERZEGOVINA, Country.MONTENEGRO),
+    officialTerritories = listOf(Country.SERBIA),
+    otherTerritories = listOf(Country.CROATIA, Country.BOSNIA_AND_HERZEGOVINA, Country.MONTENEGRO),
   ),
   SLOVAK(
     code = "sk",
-    officialCountries = listOf(Country.SLOVAKIA),
-    otherCountries = listOf(Country.SERBIA, Country.CZECHIA, Country.HUNGARY, Country.POLAND, Country.ROMANIA, Country.UKRAINE),
+    officialTerritories = listOf(Country.SLOVAKIA),
+    otherTerritories = listOf(Country.SERBIA, Country.CZECHIA, Country.HUNGARY, Country.POLAND, Country.ROMANIA, Country.UKRAINE),
   ),
   SLOVENIAN(
     code = "sl",
-    officialCountries = listOf(Country.SLOVENIA),
-    otherCountries = listOf(Country.AUSTRIA, Country.ITALY),
+    officialTerritories = listOf(Country.SLOVENIA),
+    otherTerritories = listOf(Country.AUSTRIA, Country.ITALY),
   ),
   SPANISH(
     code = "es",
-    officialCountries = listOf(Country.SPAIN, Country.MEXICO, Country.COLOMBIA, Country.ARGENTINA, Country.CHILE, Country.NICARAGUA, Country.URUGUAY, Country.EQUATORIAL_GUINEA, Country.PANAMA, Country.COSTA_RICA, Country.EL_SALVADOR, Country.PARAGUAY, Country.HONDURAS, Country.DOMINICAN_REPUBLIC, Country.CUBA, Country.BOLIVIA, Country.ECUADOR, Country.GUATEMALA, Country.VENEZUELA, Country.PERU, Country.PARAGUAY, Country.BOLIVIA),
-    otherCountries = listOf(Country.USA, Country.BELIZE, Country.ANDORRA, Country.GIBRALTAR),
+    officialTerritories = listOf(Country.SPAIN, Country.MEXICO, Country.COLOMBIA, Country.ARGENTINA, Country.CHILE, Country.NICARAGUA, Country.URUGUAY, Country.EQUATORIAL_GUINEA, Country.PANAMA, Country.COSTA_RICA, Country.EL_SALVADOR, Country.PARAGUAY, Country.HONDURAS, Country.DOMINICAN_REPUBLIC, Country.CUBA, Country.BOLIVIA, Country.ECUADOR, Country.GUATEMALA, Country.VENEZUELA, Country.PERU, Country.PARAGUAY, Country.BOLIVIA, Region.CANARY_ISLANDS, Region.LATIN_AMERICA),
+    otherTerritories = listOf(Country.USA, Country.BELIZE, Country.ANDORRA, Country.GIBRALTAR),
   ),
   SWEDISH(
     code = "sv",
-    officialCountries = listOf(Country.SWEDEN, Country.FINLAND),
-    otherCountries = listOf(Country.ESTONIA, Country.LATVIA),
+    officialTerritories = listOf(Country.SWEDEN, Country.FINLAND),
+    otherTerritories = listOf(Country.ESTONIA, Country.LATVIA),
   ),
   TAMIL(
     code = "ta",
-    officialCountries = listOf(Country.INDIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.INDIA),
+    otherTerritories = emptyList(),
   ),
   TAGALOG(
     code = "tl",
-    officialCountries = listOf(Country.PHILIPPINES),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.PHILIPPINES),
+    otherTerritories = emptyList(),
   ),
   TELUGU(
     code = "te",
-    officialCountries = listOf(Country.INDIA),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.INDIA),
+    otherTerritories = emptyList(),
   ),
   THAI(
     code = "th",
-    officialCountries = listOf(Country.THAILAND),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.THAILAND),
+    otherTerritories = emptyList(),
   ),
   TURKISH(
     code = "tr",
-    officialCountries = listOf(Country.TURKEY),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.TURKEY),
+    otherTerritories = emptyList(),
   ),
   UKRAINIAN(
     code = "uk",
-    officialCountries = listOf(Country.UKRAINE),
-    otherCountries = listOf(Country.KAZAKHSTAN, Country.MOLDOVA, Country.POLAND, Country.ROMANIA, Country.LITHUANIA, Country.SLOVAKIA),
+    officialTerritories = listOf(Country.UKRAINE),
+    otherTerritories = listOf(Country.KAZAKHSTAN, Country.MOLDOVA, Country.POLAND, Country.ROMANIA, Country.LITHUANIA, Country.SLOVAKIA),
   ),
   URDU(
     code = "ur",
-    officialCountries = listOf(Country.PAKISTAN),
-    otherCountries = listOf(),
+    officialTerritories = listOf(Country.PAKISTAN),
+    otherTerritories = listOf(),
   ),
   UZBEK(
     code = "uz",
-    officialCountries = listOf(Country.UZBEKISTAN),
-    otherCountries = listOf(Country.KAZAKHSTAN, Country.KYRGYZSTAN, Country.TAJIKISTAN, Country.TURKMENISTAN),
+    officialTerritories = listOf(Country.UZBEKISTAN),
+    otherTerritories = listOf(Country.KAZAKHSTAN, Country.KYRGYZSTAN, Country.TAJIKISTAN, Country.TURKMENISTAN),
   ),
   VIETNAMESE(
     code = "vi",
-    officialCountries = listOf(Country.VIETNAM),
-    otherCountries = emptyList(),
+    officialTerritories = listOf(Country.VIETNAM),
+    otherTerritories = emptyList(),
   ),
   ;
 
-  val defaultCountry = officialCountries.first()
+  val defaultCountry = officialTerritories.first()
+
+  fun displayName() = commonDisplayName()
 
   companion object {
-    fun from(identifier: String): Language =
-      requireNotNull(fromOrNull(identifier)) { "Can't get language from $identifier" }
+    fun from(identifier: String): Language = requireNotNull(fromOrNull(identifier)) { "Can't get language from $identifier" }
 
-    fun fromOrNull(identifier: String?): Language? =
-      values().firstOrNull { it.code.equals(identifier, ignoreCase = true) || it.legacyCode?.equals(identifier, ignoreCase = true) == true }
+    fun fromOrNull(identifier: String?): Language? = entries.firstOrNull { it.code.equals(identifier, ignoreCase = true) || it.legacyCode?.equals(identifier, ignoreCase = true) == true }
 
-    fun fromLocale(locale: String): Language =
-      requireNotNull(fromLocaleOrNull(locale)) { "Can't get language from $locale" }
+    fun fromLocale(locale: String): Language = requireNotNull(fromLocaleOrNull(locale)) { "Can't get language from $locale" }
 
-    fun fromLocaleOrNull(locale: String?): Language? =
-      fromOrNull(localeSplit(locale).getOrNull(0))
+    fun fromLocaleOrNull(locale: String?): Language? = fromOrNull(localeSplit(locale).getOrNull(0))
   }
 }
 
 /** Returns the display name of [Language] in the current language. */
-expect fun Language.displayName(): String
+internal expect fun Language.commonDisplayName(): String